refactor(e2e): move the ledger sync layer to live-e2e-shared
