+++ libs/live-e2e-shared/package.json
@@ -17,11 +17,14 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@domain/entity-currency-crypto": "workspace:*",
+    "@features/platform-wallet-sync": "workspace:^",
     "@ledgerhq/device-core": "workspace:^",
     "@ledgerhq/device-management-kit": "catalog:",
     "@ledgerhq/devices": "workspace:*",
+    "@ledgerhq/ledger-key-ring-protocol": "workspace:^",
     "@ledgerhq/ledger-wallet-framework": "workspace:^",
     "@ledgerhq/live-common": "workspace:^",
+    "@ledgerhq/live-wallet": "workspace:^",
     "@ledgerhq/live-dmk-speculos": "workspace:^",
     "@ledgerhq/live-signer-aleo": "workspace:^",
     "@ledgerhq/live-signer-evm": "workspace:^",
@@ -30,6 +33,7 @@
     "@ledgerhq/speculos-transport": "workspace:^",
     "@ledgerhq/types-devices": "workspace:^",
     "@ledgerhq/types-live": "workspace:^",
+    "@shared/cloud-sync": "workspace:^",
     "@shared/env": "workspace:*",
     "@shared/feature-flags": "workspace:*",
     "axios": "catalog:",
